Fixed 'price' variable type to match other parts of project

+++ Engine/Weapon.cs
@@ -12,7 +12,7 @@ namespace Engine
 		public int MaximumDamage { get; set; }
 
 		public Weapon(int id, string name, string namePlural, int minimumDamage,
-			int maximumDamage, double price) : base(id, name, namePlural, price)
+			int maximumDamage, int price) : base(id, name, namePlural, price)
 		{
 			MinimumDamage = minimumDamage;
 			MaximumDamage = maximumDamage;
